requirements.txt
allure_python_commons==2.9.45
allure-pytest==2.9.45
ddddocr==1.4.3
opencv_python==4.5.5.64
openpyxl==3.0.10
Pillow==9.1.1
pytesseract==0.3.9
pytest==7.1.2
PyYAML==6.0
requests==2.27.1
pytest-ordering==0.6
pytest-dependency==0.5.1
pytest-xdist==2.5.0
pytest-rerunfailures==10.2
pytest-repeat==0.9.1
selenium==4.3.0
paramiko==2.11.0
jsonpath==0.82
Faker==13.14.0

